% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bc_files.R
\name{read_bc_files}
\alias{read_bc_files}
\title{Load and merge into single data frame BC files}
\usage{
read_bc_files(bc_folder_path, bc_columns = TRUE, hour = FALSE)
}
\arguments{
\item{bc_folder_path}{Upper level folder with BC data.}

\item{bc_columns}{Bool. If TRUE, returns data.frame with BC columns.
BC1-7 are in ug/m3 and BB column in (\%).}

\item{hour}{Bool. If TRUE, it returns hourly averages.}
}
\value{
data.frame
}
\description{
Load and merge into single data frame BC files
}
\examples{
\dontrun{
# Load data for 2022
bc_dir <- "~/lumiar_data/BC/2022/"
bc_22 <- read_bc_files(bc_dir)
}
}
