% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gases_files.R
\name{read_gases_files}
\alias{read_gases_files}
\title{Load and merge single data frame XLS gases file.}
\usage{
read_gases_files(gases_folder_path, pattern)
}
\arguments{
\item{gases_folder_path}{Folder with the xls files.}

\item{pattern}{Pattern to read the xls files. "Hour", "15 Minutes",
"Minutes".}
}
\value{
data.frame
}
\description{
Load and merge single data frame XLS gases file.
}
\examples{
\dontrun{
# Load march hourly data files
gases_folder <- "/home/myUser/Data/gases/march"
gases <- read_gases_files(gases_folder, "Hour")
}
}
