% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lumasense_files.R
\name{read_lumasense_files}
\alias{read_lumasense_files}
\title{Load and merge into single data frame Lumasense files}
\usage{
read_lumasense_files(lumasense_folder_path, pol_columns = TRUE, hour = FALSE)
}
\arguments{
\item{lumasense_folder_path}{Folder were Lumasense files are located.}

\item{pol_columns}{Bool. If TRUE, it returns CO, CO2, and C2H5OH.}

\item{hour}{Bool. If TRUE, it returns hourly averages.}
}
\value{
data.frame
}
\description{
Load and merge into single data frame Lumasense files
}
\examples{
\dontrun{
# Load data for 2022
luma_dir <- "~/lumiar_data/BC/2022/"
luma_22 <- read_lumasense_files(bc_dir)
}
}
