% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nh3_files.R
\name{read_nh3_files}
\alias{read_nh3_files}
\title{Load and merge into single data frame NH3 files}
\usage{
read_nh3_files(
  nh3_folder_path,
  to_local = TRUE,
  hour = FALSE,
  col_names = c("DATE", "TIME", "NH3_30s")
)
}
\arguments{
\item{nh3_folder_path}{Upper level folder path with month and days folder.}

\item{to_local}{Change to America/Sao_Paulo time zone.}

\item{hour}{Bool. If TRUE, It returns hourly averages.}

\item{col_names}{Columns to retrieve from raw data.}
}
\value{
data.frame
}
\description{
Load into a single data frames the hour files from the NH3 analyzer.
}
\examples{
\dontrun{
# Load data from December 2022
nh3_dic_path <- "~/lumiar_data/2022/12/"
nh3_dic22 <- read_nh3_files(nh3_dic_path)
}
}
